+++ Readme.md
@@ -47,3 +47,20 @@ git clone SSH-адрес_вашего_форка
 <a href="https://htmlacademy.ru/intensive/react"><img align="left" width="50" height="50" title="HTML Academy" src="https://up.htmlacademy.ru/static/img/intensive/react/logo-for-github.png"></a>
 
 Репозиторий создан для обучения на профессиональном онлайн‑курсе «[React. Разработка сложных клиентских приложений](https://htmlacademy.ru/intensive/react)» от [HTML Academy](https://htmlacademy.ru).
+
+## Домашние задания
+
+#### 3.13 Маршрут по тайным тропам
+
+1. Создайте отдельный компонент для страницы «404». Дизайн остаётся на ваше усмотрение. В самом простом случае, это может быть страница с текстом «404 Not Found» и ссылкой для перехода на главную страницу приложения. На эту страницу пользователь будет перенаправлен в случае обращения к несуществующей странице (например, через адресную строку).
+
+2. Импортируйте все необходимые компоненты из пакета react-router-dom (BrowserRouter, Route, Routes).
+
+3. С помощью перечисленных ниже компонентов опишите все маршруты в приложении. По каждому адресу должен отрисовываться соответствующий компонент страницы.
+   Main /
+   Login /login
+   Favorites /favorites
+   Offer /offer/:id
+   Предусмотрите маршрут для обработки ситуаций, когда пользователь обращается к несуществующей странице (например, с помощью адресной строки). В этом случае он перенаправляется на страницу «404».
+
+4. Создайте компонент для описания приватных маршрутов. Этот компонент должен перенаправлять пользователя на страницу «Login» если он не авторизован. Воспользуйтесь новым компонентом и обновите маршрут для страницы «Favorites». Доступ к этой странице доступен только для авторизованных пользователей. Пока мы не взаимодействуем с настоящим сервером, передавайте компоненту информацию, что пользователь всегда не авторизован. Полноценную реализацию сделаем позже.
