+++ src/components/PrivateRoute/index.tsx
@@ -1,0 +1,18 @@
+import { Navigate } from 'react-router-dom';
+import { AppRoute, AuthorizationStatus } from '../../const';
+
+type PrivateRouteProps = {
+  authorizationStatus: AuthorizationStatus;
+  children: React.ReactNode;
+};
+
+//компонент для описания приватных маршрутов
+function PrivateRoute({ authorizationStatus, children }: PrivateRouteProps) {
+  return authorizationStatus === AuthorizationStatus.Auth ? (
+    children
+  ) : (
+    <Navigate to={AppRoute.Login} />
+  );
+}
+
+export default PrivateRoute;
