+++ src/const.ts
@@ -1,0 +1,12 @@
+export enum AppRoute {
+    Login = '/login',
+    Favorites = '/favorites',
+    Offer = '/offer',
+    Main = '/'
+}
+
+export enum AuthorizationStatus {
+    Auth = 'AUTH',
+    NoAuth = 'NO_AUTH',
+    Unknown = 'UNKNOWN',
+}
