+++ src/pages/Main/Content.tsx
@@ -1,0 +1,87 @@
+import { Place } from '../../App';
+import Header from '../../components/Header';
+import PlaceCard from '../../components/PlaceCard';
+import Location from '../../components/Location';
+import { Helmet } from 'react-helmet-async';
+
+const cities: string[] = [
+  'Paris',
+  'Cologne',
+  'Brussels',
+  'Amsterdam',
+  'Hamburg',
+  'Dusseldorf',
+];
+
+type MainProps = {
+  places: Place[];
+};
+
+function Main({ places }: MainProps): React.ReactNode {
+  return (
+    <>
+      <Helmet>
+        <title>6 cities</title>
+      </Helmet>
+      <Header />
+      <div className="page page--gray page--main">
+        <main className="page__main page__main--index">
+          <h1 className="visually-hidden">Cities</h1>
+          <div className="tabs">
+            <section className="locations container">
+              <ul className="locations__list tabs__list">
+                {cities.map((item, index) => (
+                  <Location city={item} key={index} />
+                ))}
+              </ul>
+            </section>
+          </div>
+          <div className="cities">
+            <div className="cities__places-container container">
+              <section className="cities__places places">
+                <h2 className="visually-hidden">Places</h2>
+                <b className="places__found">312 places to stay in Amsterdam</b>
+                <form className="places__sorting" action="#" method="get">
+                  <span className="places__sorting-caption">Sort by</span>
+                  <span className="places__sorting-type" tabIndex={0}>
+                    Popular
+                    <svg className="places__sorting-arrow" width="7" height="4">
+                      <use xlinkHref="#icon-arrow-select"></use>
+                    </svg>
+                  </span>
+                  <ul className="places__options places__options--custom places__options--opened">
+                    <li
+                      className="places__option places__option--active"
+                      tabIndex={0}
+                    >
+                      Popular
+                    </li>
+                    <li className="places__option" tabIndex={0}>
+                      Price: low to high
+                    </li>
+                    <li className="places__option" tabIndex={0}>
+                      Price: high to low
+                    </li>
+                    <li className="places__option" tabIndex={0}>
+                      Top rated first
+                    </li>
+                  </ul>
+                </form>
+                <div className="cities__places-list places__list tabs__content">
+                  {places.map((place) => (
+                    <PlaceCard place={place} key={place.id} />
+                  ))}
+                </div>
+              </section>
+              <div className="cities__right-section">
+                <section className="cities__map map"></section>
+              </div>
+            </div>
+          </div>
+        </main>
+      </div>
+    </>
+  );
+}
+
+export default Main;
