+++ src/pages/NotFound/index.tsx
@@ -1,0 +1,7 @@
+import NotFound from './Content';
+
+function NotFoundPage(): React.ReactNode {
+  return <NotFound />;
+}
+
+export default NotFoundPage;
